PBC-aware rebuild detection, optional position wrapping, misc (#5)
* fix ewald_real_space mask_value default to num_atoms


* fix docstrings in rebuild_detection and JAX neighbor list API


* fix JAX neighbor list dispatch to match torch dispatch


* feat(neighbors): make position wrapping optional in naive neighbor list

Add wrap_positions bool parameter (default True) to all naive neighbor
list PBC functions across warp, PyTorch, and JAX layers. When False,
skip compute_inv_cells and wrap_positions kernels, saving two GPU
launches per call for pre-wrapped positions.

Update neighborlist user guide and FAQ


* Fix MIC rebuild detection, LJ allocation, and test warnings

- Add PBC-aware rebuild detection kernels and bindings (Warp, Torch, JAX)
- Default max_nbins is 8192
- Pass cell_inv to rebuild check in MD benchmark/example infra to use
  minimum-image convention, fixing spurious rebuilds at periodic boundaries
- Sync wp_cell_inv in NPT/NPH force callbacks after cell updates
- Add pre-allocated output buffers (energies_out, forces_out, virial_out)
  to lj_energy_forces to avoid per-step allocation
- Fix JAX int64-to-int32 scatter FutureWarning across 7 neighbor files

* refactor(neighbors): decode PBC shifts on-the-fly and add prewrapped fast path

Replace materialised shift buffer with on-the-fly _decode_shift_index,
fixing int32 overflow for large periodic batches (fixes #37).
Add prewrapped kernel variants that skip per_atom_cell_offsets when
wrap_positions=False, recovering the performance lost when wrapping
was introduced. Update Torch/JAX bindings, docs, benchmarks, and tests.

* fix: sync docs and examples with current public API
Remove stale references to nonexistent unbatched/batched submodules
in neighbor list docs, delete ghost multipole electrostatics section
that advertised unimplemented APIs, and fix incorrect function
signatures for compute_temperature, shake_constraints, and
rattle_constraints in the dynamics guide. Harden examples against
missing CUDA (timing helpers, JAX preflight guards, script-relative
asset paths) and fix a latent TypeError in batched temperature
computation that passed an invalid batch_idx kwarg.


---------

+++ nvalchemiops/jax/neighbors/batch_cell_list.py
@@ -357,7 +357,7 @@ def batch_build_cell_list(
     cell_offsets = jnp.concatenate(
         [
             jnp.array([0], dtype=jnp.int32),
-            jnp.cumsum(cells_per_system[:-1]),
+            jnp.cumsum(cells_per_system[:-1], dtype=jnp.int32),
         ]
     )
 
@@ -378,7 +378,7 @@ def batch_build_cell_list(
     cell_atom_start_indices = jnp.concatenate(
         [
             jnp.array([0], dtype=jnp.int32),
-            jnp.cumsum(atoms_per_cell_count[:-1]),
+            jnp.cumsum(atoms_per_cell_count[:-1], dtype=jnp.int32),
         ]
     )
 
@@ -499,12 +499,12 @@ def batch_query_cell_list(
             dtype=jnp.int32,
         )
     elif rebuild_flags is None:
-        neighbor_matrix = neighbor_matrix.at[:].set(positions.shape[0])
+        neighbor_matrix = neighbor_matrix.at[:].set(jnp.int32(positions.shape[0]))
 
     if num_neighbors is None:
         num_neighbors = jnp.zeros(positions.shape[0], dtype=jnp.int32)
     elif rebuild_flags is None:
-        num_neighbors = num_neighbors.at[:].set(0)
+        num_neighbors = num_neighbors.at[:].set(jnp.int32(0))
 
     # Select kernel based on dtype
     if positions.dtype == jnp.float64:
@@ -533,7 +533,7 @@ def batch_query_cell_list(
             dtype=jnp.int32,
         )
     elif rebuild_flags is None:
-        neighbor_matrix_shifts = neighbor_matrix_shifts.at[:].set(0)
+        neighbor_matrix_shifts = neighbor_matrix_shifts.at[:].set(jnp.int32(0))
 
     if atoms_per_cell_count is None:
         max_total_cells = cell_atom_start_indices.shape[0]
